search_engine/requirements.txt
flask
python-dotenv
flask-cors
elasticsearch
requests
underthesea
# sentence-transformers
